sabs: overhaul split pkg cache logic
there's some other stuff going on here too. the cache creation is now
far more reliable. this also drops the pacman/expac dep, as we just
iterate over the $absroot to find where the package is. Somewhat reliant
on the $absroot itself being up to date, but much faster as its a series
of access(2) calls instead of a fork.